requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
python-multipart==0.0.6
fastapi-mail==1.2.8
python-dotenv~=1.0.0
boto3==1.26.158
google-cloud-storage
aiohttp~=3.8.4
sqlalchemy~=2.0.16
psycopg2-binary
alembic~=1.11.1
gspread~=5.9.0
google-auth
pydantic~=1.10.9
botocore==1.29.158
requests~=2.31.0
typing==3.7.4.3
python-magic==0.4.24
libmagic==1.0
pytz==2023.3

